read card infomation updates
