+++ minigrid/handlers.py
@@ -288,7 +288,6 @@ class CardsHandler(ReadCardBaseHandler):
         """Render the cards form."""
 
         http_protocol = 'https' if options.minigrid_https else 'http'
-
         self.render('cards.html', http_protocol=http_protocol)
 
 
@@ -356,8 +355,11 @@ class MinigridVendorsHandler(ReadCardBaseHandler):
     @tornado.web.authenticated
     def get(self, minigrid_id):
         """Render the vendors view."""
+        http_protocol = 'https' if options.minigrid_https else 'http'
+
         self.render(
             'minigrid_vendors.html',
+            http_protocol=http_protocol,
             minigrid=models.get_minigrid(self.session, minigrid_id))
 
     @tornado.web.authenticated
@@ -417,8 +419,11 @@ class MinigridCustomersHandler(ReadCardBaseHandler):
     @tornado.web.authenticated
     def get(self, minigrid_id):
         """Render the customers view."""
+        http_protocol = 'https' if options.minigrid_https else 'http'
+
         self.render(
             'minigrid_customers.html',
+            http_protocol=http_protocol,
             minigrid=models.get_minigrid(self.session, minigrid_id))
 
     @tornado.web.authenticated
@@ -482,8 +487,11 @@ class MinigridMaintenanceCardsHandler(ReadCardBaseHandler):
     @tornado.web.authenticated
     def get(self, minigrid_id):
         """Render the maintenance cards view."""
+        http_protocol = 'https' if options.minigrid_https else 'http'
+
         self.render(
             'minigrid_maintenance_cards.html',
+            http_protocol=http_protocol,
             minigrid=models.get_minigrid(self.session, minigrid_id))
 
     @tornado.web.authenticated
