+++ minigrid/templates/minigrid.html
@@ -1,31 +1,34 @@
 {% extends 'base.html' %}
 
 {% block body %}
-  {% if message is not None %}
-  <p><strong>{{ message }}</strong></p>
-  {% end %}
+  <div class="full-page-content">
+    {% if message is not None %}
+    <p><strong>{{ message }}</strong></p>
+    {% end %}
 
-  <h1>Minigrid Name: {{ minigrid.minigrid_name }}</h1>
-  <p>ID: {{ minigrid.minigrid_id }}</p>
-  <p>Error code: {{ minigrid.error_code }}</p>
-  <p>Status: {{ minigrid.status }}</p>
-  <p>Payment system ID: {{ minigrid.minigrid_payment_id }}</p>
-  <form action="/minigrids/{{ minigrid.minigrid_id }}" method="POST">
-    {% module xsrf_form_html() %}
-    <label>New payment system ID (note that this will change the encryption key): <input name="minigrid_payment_id" pattern="[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}" required title="UUID format" /></label><br>
-    <input type="submit" value="Update Payment System ID" />
-  </form>
-  <hr>
+    <h1>Minigrid Name: {{ minigrid.minigrid_name }}</h1>
+    <p>ID: {{ minigrid.minigrid_id }}</p>
+    <p>Error code: {{ minigrid.error_code }}</p>
+    <p>Status: {{ minigrid.status }}</p>
+    <p>Payment system ID: {{ minigrid.minigrid_payment_id }}</p>
+    <form action="/minigrids/{{ minigrid.minigrid_id }}" method="POST">
+      {% module xsrf_form_html() %}
+      <label>New payment system ID (note that this will change the encryption key): <input name="minigrid_payment_id" pattern="[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}" required title="UUID format" /></label><br>
+      <input type="submit" value="Update Payment System ID" />
+    </form>
+    <hr>
 
-  <p><a href="/minigrids/{{ minigrid.minigrid_id }}/vendors">Vendor list &raquo;</a></p>
-  <p><a href="/minigrids/{{ minigrid.minigrid_id }}/customers">Customer list &raquo;</a></p>
-  <p><a href="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards">Maintenance card list &raquo;</a></p>
-  <p><a href="/minigrids/{{ minigrid.minigrid_id }}/write_credit">Write credit card &raquo;</a></p>
-  <hr>
+    <p><a href="/minigrids/{{ minigrid.minigrid_id }}/vendors">Vendor list &raquo;</a></p>
+    <p><a href="/minigrids/{{ minigrid.minigrid_id }}/customers">Customer list &raquo;</a></p>
+    <p><a href="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards">Maintenance card list &raquo;</a></p>
+    <p><a href="/minigrids/{{ minigrid.minigrid_id }}/write_credit">Write credit card &raquo;</a></p>
+    <hr>
 
-  <p>Data:</p>
-  <p>TODO: What data? From where?</p>
-  <hr>
+    <p>Data:</p>
+    <p>TODO: What data? From where?</p>
+    <hr>
+
+  </div>
 
   <footer><a href="/">&laquo; Return to minigrid index</a></footer>
 {% end %}
