+++ minigrid/templates/minigrid_maintenance_cards.html
@@ -1,58 +1,67 @@
 {% extends 'base.html' %}
 
 {% block body %}
-  {% if message is not None %}
-    <p><strong>{{ message }}</strong></p>
-  {% end %}
-  <h1>Minigrid Name: {{ minigrid.minigrid_name }}</h1>
-  <p>ID: {{ minigrid.minigrid_id }}</p>
-  {% if received_info %}
-    <p>Card info:</p>
-    <p>{{ received_info }}</p>
-  {% end %}
-  <p>Maintenance Cards:</p>
-  <ul>
-    {% set zero_maintenance_cards = True %}
-    {% for maintenance_card in minigrid.maintenance_cards %}
-      {% set zero_maintenance_cards = False %}
-      <li>
-      <p>Name: {{ maintenance_card.maintenance_card_name }}</p>
-      <p>Card ID: {{ maintenance_card.maintenance_card_card_id }}</p>
-      <form action="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards" method="POST">
-        {% module xsrf_form_html() %}
-	<input type="hidden" name="action" value="write" />
-	<input type="hidden" name="maintenance_card_id" value="{{ maintenance_card.maintenance_card_id }}" />
-	{% if device_active %}
-          <input type="submit" value="Write Maintenance Card" />
-	{% else %}
-          <input type="submit" value="Write Maintenance Card" title="Touch card to reader" disabled/>
-	{% end %}
-      </form>
-      <form action="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards" method="POST">
-        {% module xsrf_form_html() %}
-	<input type="hidden" name="action" value="remove" />
-	<input type="hidden" name="maintenance_card_id" value="{{ maintenance_card.maintenance_card_id }}" />
-        <input type="submit" value="Remove Maintenance Card {{ maintenance_card.maintenance_card_name }}" />
-      </form>
-      </li>
+  <div class="full-page-content">
+    {% if message is not None %}
+      <p><strong>{{ message }}</strong></p>
     {% end %}
-  </ul>
-  {% if zero_maintenance_cards %}
-    <p>No maintenance cards</p>
-  {% end %}
-  <hr>
+    <h1>Minigrid Name: {{ minigrid.minigrid_name }}</h1>
+    <p>ID: {{ minigrid.minigrid_id }}</p>
+    <div class="card-info">
+      <div class="card-info-header">
+        <span>Card Info</span>
+      </div>
+    </div>
+    <p>Maintenance Cards:</p>
+    <ul>
+      {% set zero_maintenance_cards = True %}
+      {% for maintenance_card in minigrid.maintenance_cards %}
+        {% set zero_maintenance_cards = False %}
+        <li>
+        <p>Name: {{ maintenance_card.maintenance_card_name }}</p>
+        <p>Card ID: {{ maintenance_card.maintenance_card_card_id }}</p>
+        <form action="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards" method="POST">
+          {% module xsrf_form_html() %}
+  	<input type="hidden" name="action" value="write" />
+  	<input type="hidden" name="maintenance_card_id" value="{{ maintenance_card.maintenance_card_id }}" />
+  	{% if device_active %}
+            <input type="submit" value="Write Maintenance Card" />
+  	{% else %}
+            <input type="submit" value="Write Maintenance Card" title="Touch card to reader" disabled/>
+  	{% end %}
+        </form>
+        <form action="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards" method="POST">
+          {% module xsrf_form_html() %}
+  	<input type="hidden" name="action" value="remove" />
+  	<input type="hidden" name="maintenance_card_id" value="{{ maintenance_card.maintenance_card_id }}" />
+          <input type="submit" value="Remove Maintenance Card {{ maintenance_card.maintenance_card_name }}" />
+        </form>
+        </li>
+      {% end %}
+    </ul>
+    {% if zero_maintenance_cards %}
+      <p>No maintenance cards</p>
+    {% end %}
+    <hr>
 
-  <p>Add maintenance card:</p>
-  <form action="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards?action=create" method="POST">
-    {% module xsrf_form_html() %}
-    <label>Name: <input name="maintenance_card_name" required /></label><br>
-    <label>Card ID (0000-9999): <input name="maintenance_card_card_id" pattern="\d{4}" required /></label><br>
-    <input type="submit" value="Add Maintenance Card" />
-  </form>
-  <hr>
+    <p>Add maintenance card:</p>
+    <form action="/minigrids/{{ minigrid.minigrid_id }}/maintenance_cards?action=create" method="POST">
+      {% module xsrf_form_html() %}
+      <label>Name: <input name="maintenance_card_name" required /></label><br>
+      <label>Card ID (0000-9999): <input name="maintenance_card_card_id" pattern="\d{4}" required /></label><br>
+      <input type="submit" value="Add Maintenance Card" />
+    </form>
+    <hr>
+
+  </div>
 
   <footer>
     <p><a href="/minigrids/{{ minigrid.minigrid_id }}">&laquo; Return to minigrid {{ minigrid.minigrid_name }}</a></p>
     <p><a href="/">&laquo; Return to minigrid index</a></p>
   </footer>
+
+  <script>
+    var http_protocol = '{{ http_protocol }}';
+  </script>
+  <script type="text/javascript" src="{{ static_url('dist/read_card.bundle.js') }}"></script>
 {% end %}
